Add read functionality from firestore db
