+++ data/colors.js
@@ -1,0 +1,37 @@
+const dateOptions = { year: "numeric", month: "long", day: "numeric" };
+const colors = [
+  {
+    productName: "tree-runners",
+    colors: [
+      {
+        colorId: "mist",
+        colorName: "Mist (White Sole)",
+        media: [
+          "../Images/products/tree-runners/mist1.avif",
+          "../Images/products/tree-runners/mist2.avif",
+          "../Images/products/tree-runners/mist3.avif",
+        ],
+      },
+      {
+        colorId: "buoyant-blue",
+        colorName: "Buoyant Blue (Blizzard Sole)",
+        media: [
+          "../Images/products/tree-runners/buoyant-blue1.avif",
+          "../Images/products/tree-runners/buoyant-blue2.avif",
+          "../Images/products/tree-runners/buoyant-blue3.avif",
+        ],
+      },
+      {
+        colorId: "wasatch",
+        colorName: "Wasatch (White Sole)",
+        media: [
+          "../Images/products/tree-runners/wasatch1.avif",
+          "../Images/products/tree-runners/wasatch2.avif",
+          "../Images/products/tree-runners/wasatch3.avif",
+        ],
+      },
+    ],
+  },
+];
+
+export default colors;
