+++ products/tree-runners.html
@@ -375,6 +375,5 @@
       </div>
     </section>
     <script type="module" src="../scripts/Product/index.js"></script>
-    <script type="module" src="../scripts/firebase.js"></script>
   </body>
 </html>
